feat: migrate from sqlite to postgres
Sqlc didn't work well with sqlite especially when using array parameters
and dynamic fitlering resulting in strange bugs in code generation.

+++ internal/security/security.go
@@ -23,7 +23,7 @@ const (
 type Session struct {
 	id string
 
-	UserID   int64
+	UserID   int32
 	Username string
 	Role     Role
 }
